+++ config/initializers/i18n.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-# we allow the Zulu locale (zu) as it is used for In-Context translation
-# @see https://support.crowdin.com/in-context-localization/
-I18n.config.available_locales = :zu