+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Changed some default values for new installations
 - Updated documentation
 - Fix a bug: installation without nginx does not remove the service from the docker-compose file
+- Fix a bug: default twitter feed is invalid
 
 ## v4.3.3 2020 April 1st