+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix a bug: default twitter feed is invalid
 - Fix a bug: default nginx configuration does not allows secure cookies
 - Fix a bug: in-context translation is loading invalid locale for MessageFormat
+- Fix a bug: invalid link to upgrade procedure
 
 ## v4.3.3 2020 April 1st