+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Compatible database with Fab-manager v1, to allow upgrades
 - Updated documentation
 - Changed In-Context pseudo-language to Zulu instead of Acholi
+- Allow removing contacts from the about page
 - Fix a bug: installation without nginx does not remove the service from the docker-compose file
 - Fix a bug: default twitter feed is invalid
 - Fix a bug: default nginx configuration does not allows secure cookies